Appearance upgrade
The text for the card information form was in need of a revamp so I changed the text color and font to make it more attractive. I also found a logo that somewhat fits the project name and implemented that so it's more attractive when working with it in a testing environment.